# api/docs/swagger.yaml
basePath: /v1
definitions:
  api.Response:
    properties:
      message:
        type: string
    type: object
  model.Diagnostic:
    properties:
      dateTime:
        type: string
      diagnostic:
        minLength: 10
        type: string
      id:
        type: integer
      patientID:
        type: integer
      prescription:
        $ref: '#/definitions/model.Prescription'
    required:
    - dateTime
    - diagnostic
    - patientID
    type: object
  model.DiagnosticRequest:
    properties:
      dateTime:
        type: string
      diagnostic:
        minLength: 10
        type: string
      patientID:
        type: integer
      prescription:
        $ref: '#/definitions/model.PrescriptionRequest'
    required:
    - dateTime
    - diagnostic
    - patientID
    type: object
  model.DiagnosticWithPatient:
    properties:
      dateTime:
        type: string
      diagnostic:
        type: string
      id:
        type: integer
      patient:
        $ref: '#/definitions/model.Patient'
      prescription:
        $ref: '#/definitions/model.Prescription'
    type: object
  model.Patient:
    properties:
      documentID:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      phone:
        type: string
    type: object
  model.Prescription:
    properties:
      description:
        type: string
      diagnosticID:
        type: integer
      dueDate:
        type: string
      id:
        type: integer
    type: object
  model.PrescriptionRequest:
    properties:
      description:
        type: string
      dueDate:
        type: string
    type: object
info:
  contact: {}
  description: A Healthcare service API in Go using Gin
  title: Healthcare API
  version: "1.0"
paths:
  /diagnostics:
    get:
      description: List dignostic data from bd.
      parameters:
      - description: Filter diagnostics by firstName
        in: query
        name: firstName
        type: string
      - description: Filter diagnostics by lastName
        in: query
        name: lastName
        type: string
      - description: Filter diagnostics by date
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.DiagnosticWithPatient'
            type: array
      summary: List diagnostics
      tags:
      - diagnostics
    post:
      description: Save diagnostic data in bd.
      parameters:
      - description: Create diagnostic
        in: body
        name: diagnostic
        required: true
        schema:
          $ref: '#/definitions/model.DiagnosticRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Diagnostic'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
      summary: Create diagnostic
      tags:
      - diagnostics
swagger: "2.0"